Correct wrong extract class
