+++ src/Tasks/Microsoft.NET.Build.Tasks/ResolveApphostShimFromFrameworkReferences.cs
@@ -3,9 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using Microsoft.Build.Framework;
-using static Microsoft.NET.Build.Tasks.ResolveFrameworkReferences;
 
 namespace Microsoft.NET.Build.Tasks
 {
